Prevent race condition on deep post paths
fs::create_dir_all is racy and will not do well when run in parallel for
a number of sufficiently deep directories with same parents.

I fixed it by writing my own version that is thread-safe.

I also reorganized some of the document code, it seems more clean to me
now at least.