Implement custom document paths (permalinks) (#107)
* Move parse_document code into Document

* Prevent race condition on deep post paths

fs::create_dir_all is racy and will not do well when run in parallel for
a number of sufficiently deep directories with same parents.

I fixed it by writing my own version that is thread-safe.

I also reorganized some of the document code, it seems more clean to me
now at least.

* Implement custom document paths

Some things still need to be finished, like proper slugification. But
that should be for follow up bugs.

Note that this also removes the automatically injected variable `name`
which used to be the file name without the extension. I think it never
really made sense to have it.