Error handling and index.html search
As suggested in the pull request discussion, I moved the handle to a
separate function and matched all errors instead of unwrapping
everything.

This also makes the handler search for a "index.html" file whenever
possible. This means that "/blog/engineer" (or any path, really) will
try to point to "/blog/engineer/index.html", unless this path
corresponds to a file (see Rust documentation for PathBuf::is_file() for
more details).