Better error handling and code improvements
Grab the path from the `req.uri` instead of relying in the `format!`
macro.

Changed some `match`es to `if-let`s when dealing with only the error.

Removed unnecessary `path.clone()` call when constructing serve_path.

If the page is not found, a simple 404 page is displayed, and the
NotFound status code is sent.