Render posts before documents, remove threading
In #138 and #145 I noticed that it might be beneficial to render posts
before documents, because documents might want to access post
information such as content or excerpt that requires all posts to be
rendered.

This also removes the admittedly naive approach of spawning a thread per
document which at least by my measurements adds way too much overhead to
be of any use. I will open a new issue for experimenting with ways to
add good parallelism.

Fixes #138 by adding a "content" attribute to posts.