feat(attributes): Add input-file based attributes
Add document attributes "title", "slug", and "ext" if needed based on the
input-file.

The primary motivation for this change is to allow a custom "path" while
retaining the original :slug. A more surgical approach was taken rather
than trying to achieve 100% Jekyll-compatibility though Jekyll does serve
as an inspiration for how these features work.

The "slug" attribute defaults to a lower-case slugified version of the
document's filename.  The primary use is for specifying a custom "path"
attribute that still includes the original file name (via :slug).

Jekyll compatibility:
- Precedence: Allows a user to override "slug" in their front matter.
- Template variable for path: The default slug attempts to be formatted
  like Jekyll
Jekyll incompatibility:
- General: Jekyll excludes "YYYY-MM-DD-" from the slug. Cobalt would need
  additional work done to make "date" more Jekyll compatible before this
  is worth implementing.
- Template variable for path: An overridden slug is not slugified in
  Cobalt. This is in part for simplicity and in part trusting the user.
- Attribute: Jekyll does not sluggify the attribute but instead expects
  the user to slugify it with a filter.

The "title" attribute is now defaulted to a title-cased, slugified version
of the document's filename. The primary use is to minimize user's having
to repeat themselves in their liquid document.

Jekyll compatibility:
- Precedence: Allows a user to override "title" in their front matter.
  Does not use the overridden "slug" attribute if provided.
Jekyll incompatibility:
- General: Jekyll excludes "YYYY-MM-DD-" from the title.
- Template variable for path: :title is supposed to be a "pretty" mixed
  case version of :slug
- Attribute: Jekyll does not slugify the file name before title-casing it.

The "ext" attribute defaults to the file extension of the liquid document
("md", "liquid", etc)

Jekyll compatibility:
- Precedence: Allows a user to override "ext" in their front matter.
Jekyll incompatibility:
- Attribute: Jekyll includes the ".".

The :name template variable for path (slugified file_stem, not overriding
allowed) was also going to be added but for some reason it caused a test
to fail. Since it wasn't needed for the primary purpose of this change, it
was skipped rather than time being spent on investigating the failure.