refactor default templates directory to '_templates'
The name `_layouts` is a bit confusing, as what really exists in the
`_layouts` directory is file templates, to be used in pages or posts.
This patch refactors `Config.layouts` to `Config.templates`, and all
reference values from `_layouts` to `_templates`.

Additionally, extraneous backticks (in the README) for inline
preformatted text were fixed up to use the minimal requirement of a
single backtick.