Add include option to config
Override usually ignored folders/files to be included. Such as those
prefixed with `.` or `_`. Related to #221. Sometimes hidden folders or
files need to be included for example when verifying keybase identities,
or when getting Let's Encrypt certificates, which have files in a
`.well-known` folder.

To include specific directories or files add something like the
following to .cobalt.yml

    include:
      - .well-known

This will include the `.well-known` directory in the build

`exclude` will take precedence over `include`. Having matching entries
in either list will exclude over include. It is however possible to do
something like this:

    include:
        - .adirectory
    exclude:
        - .adirectory/afile

to include the `.adirectory` without `afile` in it