fix: Align `FilesIterator` with `includes_file`
`FilesIterator` tests the root and then tests the child directories.
Originally, `includes_file` was testing only the leaf.  This change
attempts to simulate walking the directory tree.

In the short term, this should make the unit tests more relevant.  In
the long term, this is fixing funtionality that will be helpful with
`cargo watch`.