chore: Explicitly generate page attributes
Previously, we dumped all of the front matter in as page attributes.
Now, native frontmatter attributes are specifically selected.

This is making it possible to move all custom attributes into their own
child Object on the page, which is important for #257.

BREAKING CHANGES

`slug`, `excerpt_separator`, `ext`, and `extends` are no longer page
attributes.  The assumption is that these weren't even really needed.