refactor: Load `.cobalt.yml` via serde
- This simplifies the config loading code.
- This is yet another stepping stone towards a revamped config.
- This allows the removal of a direct dependency.

BREAKING CHANGE:

Unknown fields are mostly not allowed now (for some reason, incorrect
syntax-highlight fields are).  There isn't really a use case for unknown
fields but room for a lot of confusion (typos in fields leading to odd
behavior).

`syntax-highlight` has been renamed to `syntax_highlight`.  This is more
consistent with the rest of the file.