fix(jekyll): Convert to external, rather than internal, format
Currently, cobalt has two different formats, frontmatter and
legacy/wildwest.  The decoupling is meant to allow us to experiment with
what we should make the public format be without breaking people on
every little change.  When we're done, we'll create an upgrade tool and
switch.

The jekyll converter was applied to the internal format rather than the
public format.