fix(watch): Rebuilding ignores dest
The previous work for defining paths probably broke
things because it was operating on an absolute path rather than a relative path.

In addition, the previous fix was a bit awkward because it was using `dest` relative
to root, rather than making `dest` relative to `source`.

This should address those problems and clean up the code while at it.