fix: Auto-ignore dest in more cases
In manual testing, I found the current setup could not handle
```
source: "."
dest: "./build"
```

Ideally, we'd canonicalize these paths but at least dest might not exist
yet, so it'd fail.  I hand-implemented some path normalization.  I now
realize a real solution could possibly be made with `Path::components`.
That can be an exercise for if we run into more cases of this breaking.