feat(front): Set published_date from filename
This is inspired by #197 and Jekyll.

Jekyll gets the filename from a file.  The main difference between
Jekyll and this change is that we aren't supporting 2-digit years.

In #197, the idea was brought up of extracting the date from the
directory structure.  I like this approach.  I could imagine having at
least year directories (don't blog enough to make day directories worth
while).  The challenge comes in how to detect this.