fix(CI): Try to fix coverage reporting
I think its because we are using an old cargo-travis that doesn't
understand new cargo features.

I had to disable the caching of cargo-travis to upgrade it because we
have no way to make the cache auto-invalidate.  We'll just have to wait
until we can assume its propogate to all travis caches.