feat(front): Standardize permalink behavior
Two main aspects
- Collections are now slugified
- Custom variables are now accessed as `:data.<VAR>` to avoid collisions
  with future built-in variables

The challenge with writing the conversion is we don't know where a
`:data.<VAR>` field ends.  Originally, it was written to ignore
`:data.<VAR>` variables.  Instead it consumes everything until the next
`:variable` as part of `:data.<VAR>`.  This isn't noticed in this change
but will become obvious when changing the format.

Part of the work for #198

BREAKING CHANGE:
- permalinks with `:collection` might move due to slugification
- Custom variables need to be switched to `:data.<VAR>`. `cobalt
  migrate` is unsufficient for this, instead it will remove the troubled
  permalink

To be clear, this was tested on 10 sites and broke none of them.