feat(front): Change permalink to well-defined format
The original jekyll-inspired format is not well defined.  The only way
to parse a variable is to match it against all know variables because
there is no end-token.  This can also lead to ambiguity if one variable
name is a subset of another.

This switches the format to use liquid.  No filters, tags, or blocks are
made available at this time but that can be re-evaluated in the future.

Part of the fix for #198

BREAKING CHANGE: permalink substitutions like `:path:output_ext` need to
be changed to `{{path}}{{output_ext}}`.

`cobalt migrate` will take care of this.  It might include too much text
in the variable but this will be caught by `cobalt build` failing.