feat(excerpt): Better define non-existent behavior
Previously, when there wasn't an exceprt (because `excerpt_separator`
was `"`), `page.excerpt` would be `""` ... unless it wasn't a post in
which it would not exist.

Now, `page.excerpt` evaluates to `Nil` for non-existent excerpts.  This
makes it easier for testing (`{% if page.excerpt %}`) and helps move
pages / posts into being treated exactly the same in how they are
processed internally ("all just collections").

Some dead code was also removed.