feat(serve): Adjust base_url for localhost
This is to make testing of sites easier.  If a site is in a child
directory on the host, the links won't work locally.  Now if someone
uses `base_url` to qualify all of the links, they'll auto-adapt between
the hosting service and `cobalt serve`.

BREAKING CHANGE: `cobalt serve` will no longer generate the exact same
page as `cobalt build`.