fix(files): sort by filename to have a reproducibly walking iterator
The underlaying structure's order is not fixed and may depend on
the filesystem implementation. Therefor the iterator needs to be
sorted by filename to guarantee a reproducible order of processing
input files.

One effect of this issue can be observed by having an unreproducible
order of posts that have the very same published_date (or none as
that falls back to default_date) purely depending on the order the
underlaying filesystem serves the files for consumption.

This fixes #457