fix: Do not render `Nil` excerpt
Setting `excerpt_separator` to the empty string is intended to disable
auto-extracted excerpts, and, as a side-effect, cause the full content
to be used in the RSS feed.

Before this change, this did not happen, as
`Document::description_to_str` would render the `Nil` value inserted
into the "excerpt" attribute. We now avoid rendering a `Nil` excerpt,
and fall back to rendering the full content instead.

Fixes: #724