chore: Fix a `match`-related beta clippy warning
Identified by beta clippy (4ee1206 2020-02-01); it remarked that
having the wildcard as an alternative is redundant, as it will always
match. Instead of getting rid of the `Some("info")` alternative, I
preserved it in order to keep all valid log level names, and added
a separate case for the fallback instead.