fix(serve): Resolve infinite loop om Windows. (#855)
Before this change, the serve command would repeatedly rebuild the website
after the first change. This was caused by the the rebuild being triggered
when the output files changed. The reason the existing check did not notice
that the change files were in the output folder is due to how Rust
canonicalizes paths on Windows. Most Windows are familar with absolute paths
that start with the drive letter. For example:

c:\windows\notepad.exe

The fully qualified version of this path, returned by canonicalize(), is:

\\?\c:\windows\notepad.exe

The destination path was not canonicalized, while the paths coming from
notify.rs were. This prevented the starts_with check from detecting that
the changed path was within the destintion folder.

I also tested this change on Linux and confirmed that it still rebuilds
when the input filess change.