feat(file-serve)!: Allow graceful shutdown of server
Assume that someone was developing an App which would serve some files temporarily, the server should be closed when it was no longer needed. But the server was just blocked for handling incoming requests, there were no valid methods to close it.

So a method for graceful shutdown of the server is introduced in this PR, also a method for checking whether a sever is running.

BREAKING CHANGE: `Server.serve()` returns `Result<(), Error>` instead of `Result<Infallible, Error>` because the server can be closed.
Remove `Debug` trait of `Server`.